+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 
 ## Next Release
+
+__New Features and Enhancements:__
+
+- Add support for `copyInstanceOnItemCopy` field for metadata templates ([#850](https://github.com/box/box-java-sdk/pull/850))
 - Add support for more fields in `BoxCollaborator.Info` ([#843](https://github.com/box/box-java-sdk/pull/843))
 
 ## 2.50.1 [2020-08-20]